+++ app/client/src/PluginActionEditor/components/PluginActionForm/hooks/useChangeActionCall.test.tsx
@@ -1,7 +1,7 @@
 import { renderHook } from "@testing-library/react-hooks/dom";
 import { useDispatch } from "react-redux";
 import { PluginType } from "entities/Action";
-import { usePluginActionContext } from "../../../PluginActionContext";
+import { usePluginActionContext } from "PluginActionEditor";
 import { changeApi, changeQuery } from "../../../store";
 import usePrevious from "utils/hooks/usePrevious";
 import { useChangeActionCall } from "./useChangeActionCall";
