+++ app/client/src/ce/pages/Applications/CreateNewAppsOption.test.tsx
@@ -59,7 +59,7 @@ const defaultStoreState = {
         config: {},
       },
     },
-    apiPane: {
+    pluginActionEditor: {
       isCreating: false,
       isRunning: {},
       isSaving: {},
