+++ app/client/src/components/editorComponents/Debugger/DebuggerTabs.tsx
@@ -20,7 +20,7 @@ import {
   DEBUGGER_LOGS,
   INSPECT_ENTITY,
 } from "ee/constants/messages";
-import { DEBUGGER_TAB_KEYS } from "./helpers";
+import { DEBUGGER_TAB_KEYS } from "./constants";
 import EntityBottomTabs from "../EntityBottomTabs";
 import { ActionExecutionResizerHeight } from "pages/Editor/APIEditor/constants";
 import { IDEBottomView, ViewHideBehaviour, ViewDisplayMode } from "IDE";
