+++ app/client/src/components/editorComponents/Debugger/constants.ts
@@ -1,0 +1,8 @@
+export enum DEBUGGER_TAB_KEYS {
+  SCHEMA_TAB = "schema",
+  RESPONSE_TAB = "response",
+  HEADER_TAB = "headers",
+  ERROR_TAB = "ERROR",
+  LOGS_TAB = "LOGS_TAB",
+  INSPECT_TAB = "INSPECT_TAB",
+}
